feat(api): add chain pipeline stage result type
