+++ api/onnx_web/chain/reduce_crop.py
@@ -6,7 +6,7 @@ from PIL import Image
 from ..params import ImageParams, Size, StageParams
 from ..server import ServerContext
 from ..worker import WorkerContext
-from .stage import BaseStage
+from .base import BaseStage
 
 logger = getLogger(__name__)
