+++ api/onnx_web/chain/source_txt2img.py
@@ -18,7 +18,7 @@ from ..diffusers.utils import (
 from ..params import ImageParams, Size, SizeChart, StageParams
 from ..server import ServerContext
 from ..worker import ProgressCallback, WorkerContext
-from .stage import BaseStage
+from .base import BaseStage
 
 logger = getLogger(__name__)
